backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==1.0.3
realtime-py==0.1.5
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
pinecone
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0